feat: enumertor service tests added
