+++ tests/ApplicationTests/Services/Common/EnumeratorServiceTests.cs
@@ -1,15 +1,7 @@
-﻿using Application.Contracts.Responses.Common;
-using Application.Errors.Services;
-using Application.Interfaces.Infrastructure.Services;
+﻿using Application.Interfaces.Infrastructure.Services;
 using Application.Services.Common;
 
 using AutoMapper;
-
-using BaseTests.Helpers;
-
-using Domain.Errors;
-
-using FluentAssertions;
 
 using Moq;
 
@@ -17,214 +9,10 @@ namespace ApplicationTests.Services.Common;
 
 [TestClass]
 [SuppressMessage("Style", "IDE0058", Justification = "Not relevant here, unit testing.")]
-public sealed class EnumeratorServiceTests : ApplicationTestBase
+public sealed partial class EnumeratorServiceTests : ApplicationTestBase
 {
 	private readonly IMapper _mapper = GetService<IMapper>();
 	private Mock<ILoggerService<EnumeratorService>> _loggerServiceMock = default!;
-
-	[TestMethod]
-	[TestCategory(nameof(EnumeratorService.GetAccountTypes))]
-	public void GetAccountTypesShouldReturnResultWhenSuccessful()
-	{
-		EnumeratorService sut = CreateMockedInstance(_mapper);
-
-		var result = sut.GetAccountTypes();
-
-		AssertionHelper.AssertInScope(() =>
-		{
-			result.Should().NotBeNull();
-			result.IsError.Should().BeFalse();
-			result.Errors.Should().BeEmpty();
-			result.Value.Should().NotBeNullOrEmpty();
-		});
-	}
-
-	[TestMethod]
-	[TestCategory(nameof(EnumeratorService.GetAccountTypes))]
-	public void GetAccountTypesShouldReturnFailedWhenExceptionGetThrown()
-	{
-		EnumeratorService sut = CreateMockedInstance(null!);
-
-		var result = sut.GetAccountTypes();
-
-		AssertionHelper.AssertInScope(() =>
-		{
-			result.Should().NotBeNull();
-			result.IsError.Should().BeTrue();
-			result.Errors.First().Should().Be(EnumeratorServiceErrors.GetCardTypesFailed);
-			result.Value.Should().BeNullOrEmpty();
-		});
-	}
-
-	[TestMethod]
-	[TestCategory(nameof(EnumeratorService.GetCardTypes))]
-	public void GetCardTypesShouldReturnResultWhenSuccessful()
-	{
-		EnumeratorService sut = CreateMockedInstance(_mapper);
-
-		ErrorOr<IEnumerable<CardTypeResponse>> result = sut.GetCardTypes();
-
-		AssertionHelper.AssertInScope(() =>
-		{
-			result.Should().NotBeNull();
-			result.IsError.Should().BeFalse();
-			result.Errors.Count.Should().Be(0);
-			result.Value.Should().NotBeNullOrEmpty();
-		});
-	}
-
-	[TestMethod]
-	[TestCategory(nameof(EnumeratorService.GetCardTypes))]
-	public void GetCardTypesShouldReturnFailedWhenExceptionGetThrown()
-	{
-		EnumeratorService sut = CreateMockedInstance(null!);
-
-		ErrorOr<IEnumerable<CardTypeResponse>> result = sut.GetCardTypes();
-
-		AssertionHelper.AssertInScope(() =>
-		{
-			result.Should().NotBeNull();
-			result.IsError.Should().BeTrue();
-			result.Errors.First().Should().Be(EnumeratorServiceErrors.GetCardTypesFailed);
-			result.Value.Should().BeNullOrEmpty();
-		});
-	}
-
-	[TestMethod]
-	[TestCategory(nameof(EnumeratorService.GetAttendanceTypes))]
-	public void GetAttendanceTypesShouldReturnResultWhenSuccessful()
-	{
-		EnumeratorService sut = CreateMockedInstance(_mapper);
-
-		ErrorOr<IEnumerable<AttendanceTypeResponse>> result = sut.GetAttendanceTypes();
-
-		AssertionHelper.AssertInScope(() =>
-		{
-			result.Should().NotBeNull();
-			result.IsError.Should().BeFalse();
-			result.Errors.Count.Should().Be(0);
-			result.Value.Should().NotBeNullOrEmpty();
-		});
-	}
-
-	[TestMethod]
-	[TestCategory(nameof(EnumeratorService.GetAttendanceTypes))]
-	public void GetAttendanceTypesShouldReturnFailedWhenExceptionGetThrown()
-	{
-		EnumeratorService sut = CreateMockedInstance(null!);
-
-		ErrorOr<IEnumerable<AttendanceTypeResponse>> result = sut.GetAttendanceTypes();
-
-		AssertionHelper.AssertInScope(() =>
-		{
-			result.Should().NotBeNull();
-			result.IsError.Should().BeTrue();
-			result.Errors.First().Should().Be(EnumeratorServiceErrors.GetAttendanceTypesFailed);
-			result.Value.Should().BeNullOrEmpty();
-		});
-	}
-
-	[TestMethod]
-	[TestCategory(nameof(EnumeratorService.GetPriorityLevelTypes))]
-	public void GetPriorityLevelTypesShouldReturnResultWhenSuccessful()
-	{
-		EnumeratorService sut = CreateMockedInstance(_mapper);
-
-		ErrorOr<IEnumerable<PriorityLevelTypeResponse>> result = sut.GetPriorityLevelTypes();
-
-		AssertionHelper.AssertInScope(() =>
-		{
-			result.Should().NotBeNull();
-			result.IsError.Should().BeFalse();
-			result.Errors.Count.Should().Be(0);
-			result.Value.Should().NotBeNullOrEmpty();
-		});
-	}
-
-	[TestMethod]
-	[TestCategory(nameof(EnumeratorService.GetPriorityLevelTypes))]
-	public void GetPriorityLevelTypesShouldReturnFailedWhenExceptionGetThrown()
-	{
-		EnumeratorService sut = CreateMockedInstance(null!);
-
-		ErrorOr<IEnumerable<PriorityLevelTypeResponse>> result = sut.GetPriorityLevelTypes();
-
-		AssertionHelper.AssertInScope(() =>
-		{
-			result.Should().NotBeNull();
-			result.IsError.Should().BeTrue();
-			result.Errors.First().Should().Be(EnumeratorServiceErrors.GetPriorityLevelTypesFailed);
-			result.Value.Should().BeNullOrEmpty();
-		});
-	}
-
-	[TestMethod]
-	[TestCategory(nameof(EnumeratorService.GetRoleTypes))]
-	public void GetRoleTypesShouldReturnResultWhenSuccessful()
-	{
-		EnumeratorService sut = CreateMockedInstance(_mapper);
-
-		ErrorOr<IEnumerable<RoleTypeResponse>> result = sut.GetRoleTypes();
-
-		AssertionHelper.AssertInScope(() =>
-		{
-			result.Should().NotBeNull();
-			result.IsError.Should().BeFalse();
-			result.Errors.Count.Should().Be(0);
-			result.Value.Should().NotBeNullOrEmpty();
-		});
-	}
-
-	[TestMethod]
-	[TestCategory(nameof(EnumeratorService.GetRoleTypes))]
-	public void GetRoleTypesShouldReturnFailedWhenExceptionGetThrown()
-	{
-		EnumeratorService sut = CreateMockedInstance(null!);
-
-		ErrorOr<IEnumerable<RoleTypeResponse>> result = sut.GetRoleTypes();
-
-		AssertionHelper.AssertInScope(() =>
-		{
-			result.Should().NotBeNull();
-			result.IsError.Should().BeTrue();
-			result.Errors.First().Should().Be(EnumeratorServiceErrors.GetRoleTypesFailed);
-			result.Value.Should().BeNullOrEmpty();
-		});
-	}
-
-	[TestMethod]
-	[TestCategory(nameof(EnumeratorService.GetWorkDayTypes))]
-	public void GetWorkDayTypesShouldReturnResultWhenSuccessful()
-	{
-		EnumeratorService sut = CreateMockedInstance(_mapper);
-
-		ErrorOr<IEnumerable<WorkDayTypeResponse>> result = sut.GetWorkDayTypes();
-
-		AssertionHelper.AssertInScope(() =>
-		{
-			result.Should().NotBeNull();
-			result.IsError.Should().BeFalse();
-			result.Errors.Count.Should().Be(0);
-			result.Value.Should().NotBeNullOrEmpty();
-		});
-	}
-
-	[TestMethod]
-	[TestCategory(nameof(EnumeratorService.GetWorkDayTypes))]
-	public void GetWorkDayTypesShouldReturnFailedWhenExceptionGetThrown()
-	{
-		EnumeratorService sut = CreateMockedInstance(null!);
-
-		ErrorOr<IEnumerable<WorkDayTypeResponse>> result = sut.GetWorkDayTypes();
-
-		AssertionHelper.AssertInScope(() =>
-		{
-			result.Should().NotBeNull();
-			result.IsError.Should().BeTrue();
-			result.Errors.First().Should().Be(EnumeratorServiceErrors.GetWorkDayTypesFailed);
-			result.Value.Should().BeNullOrEmpty();
-		});
-	}
 
 	private EnumeratorService CreateMockedInstance(IMapper mapper)
 	{
